LayoutTests/fast/table/fixed-table-layout/fixed-layout-column-colspan-wrong-size.html

<!DOCTYPE html>
<html>
<head>
<style>
table {
    width: 330px;
    table-layout: fixed;
    border-spacing: 0px;
}
td {
    padding: 0px;
}
</style>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

function test()
{
    description("Bug 20213: colspan in table-layout:fixed does not add up widths properly");

    testCell = document.getElementById("tdColspan");
    testTable = document.getElementsByTagName("table")[0];
    shouldBe("window.getComputedStyle(testCell).width", "'320px'");
    shouldBe("window.getComputedStyle(testTable).width", "'330px'");
    isSuccessfullyParsed();
}

window.addEventListener("load", test);
</script>
</head>
<body>
    <table>
        <colgroup>
            <col width="10">
            <col width="300">
            <col width="10">
            <col width="10">
        </colgroup>
        <tbody>
            <tr>
            <td></td>
            <td colspan="3" id="tdColspan"></td>
            </tr>
        </tbody>
    </table>
    <div id="description"></div>
    <div id="console"></div>
</body>
</html>
