third_party/blink/web_tests/fast/table/fixed-table-layout/fixed-layout-column-colspan-wrong-size-expected.txt

	

Bug 20213: colspan in table-layout:fixed does not add up widths properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS window.getComputedStyle(testCell).width is '320px'
PASS window.getComputedStyle(testTable).width is '330px'
PASS successfullyParsed is true

TEST COMPLETE
